+++ src/lib/index.js
@@ -3,7 +3,7 @@ import Download from './components/Download.react';
 import Lottie from './components/Lottie.react';
 import Keyboard from './components/Keyboard.react';
 import Monitor from './components/Monitor.react';
-import BurgerMenu from './components/BurgerMenu.react';
+import Burger from './components/Burger.react';
 import Ticker from './components/Ticker.react';
 import WebSocket from './components/WebSocket.react';
 import BeforeAfter from './components/BeforeAfter.react';
@@ -16,7 +16,7 @@ export {
     Lottie,
     Keyboard,
     Monitor,
-    BurgerMenu,
+    Burger,
     WebSocket,
     Ticker,
     BeforeAfter,